Basic preparations for CEGUI 0.8.
This commit cleans up the CEGUI initialisation process and hides it
completely below pVideo::Init_Video(). There’s no need to publically
expose the Init_CEGUI() function. Init_CEGUI_Data() has been removed and
incorporated into Init_CEGUI() as it makes the Init_CEGUI() function
easier to understand. The pGuiSystem and pGuiRenderer global variables
have been removed; CEGUI::System is a singleton class whose one global
instance can be retrieved at any time with
CEGUI::System::getSingleton(). The renderer is stored as a member
variable of the cVideo class, which makes it more easy to control its
storage. The destructor of cVideo is now able to take care of
uninitialisation of the CEGUI system without need from the outside.

It further changes the initialisation so that it is adapted to the new
CEGUI 0.8.x, and does likewise for the rendering process. The
introduction of CEGUI::GUIContext required a new time pulse injection,
but is certainly going to require a lot of further changes in the TSC
code. Therefore, this commit does not compile and is merely a start of
the porting effort for CEGUI 0.8.x.

Finally, it fixes #506, the logfile problem. The logfile path was never
handed to CEGUI, which couldn’t have worked. When I implemented the new
initialisation procedure for CEGUI 0.8.x, it was easy to directly do it
correctly so that the logfile problem disappears.