Delete custom CEGUI spinner subclass.
This used an undocumented CEGUI API just for the sake of getting the
value out of a spinner as a string. It shouldn’t be too hard to simply
place the return value of the regular value retrieval function in a
conversion function, that does not warrant use of undocumented APIs.