Port loading screen to CEGUI 0.8.
This commit also extracs the global Loading_Screen_*() functions into a
separate source file. As global functions they shouldn’t be “hidden”
between the ordinary member functions of cVideo.

Having the loading screen as global functions is bad code design in any
case, but it is not this branch’s task to fix that. Once we switch to a
scene-based system they are going to go away anyway.