Typedef our legacy CEGUI subclasses to the original CEGUI classes.
This is an intermediary measure to ease porting work. I don’t want to
start porting with the TSC editor, which is a very complex
tool. However, editor.hpp is included by a lot of other files and they
expect some of these classes to be available.

The legacy classes have earlier been deleted with commit
4b9ebd4ba5d4126fcf87036e241382cc910cb9ee, but not replaced with
anything, breaking compilation. The typdefs introduced now replace them
with the original classes they have been subclasses of, i.e. the
subclasses are not subclasses anymore, but typdefs of the original
classes.

This workaround obviously needs to be removed before merging the CEGUI
changes in.