Fix CEGUI eating all mouse events.
The mouse events weren't passed through to our own mouse handler,
because the invisible root window was missing the
MousePassThroughEnabled property. CEGUI consumed the mouse click events,
making clicking non-CEGUI items (like the main menu entries)
impossible.

Additionally, this commit changes the loading screen handler to not
delete the invisible root window, which is a bad idea as the properties
on it need to be carried through (for example, the above
MousePassThroughEnabled property). Incidentally, this also fixes the
segfault on exiting the game.