Catch renderer exception on editor loading.
The basics of the editor menu now appear to work. The different images
for the different menu entries are shown when the respective menu entry
is clicked. There appears to be a problem in processing the ground tags,
though, and some images throw renderer exceptions.

Also, the game now appears to need a while when exiting. This might have
to do with CEGUI cleaning up its image cache.