Sync the different places of editor enabled information.
TSC keeps the information of whether the editor is opened currently at
three different places:

* m_enabled member of cEditor class
* editor_enabled global variable
* editor_level_enabled/editor_world_enabled global variable

If these variables go out of sync weird things start to happen, because
some parts of the game think the editor is open and others do not. It
does not really help that the two global variables are set outside of
cEditor sometimes. This combination is a scholar’s example of what kind
of umainainable code results if global variables are used excessively,
and even worse if global variables are combined with the no-go of
holding the same piece of information at different places. Global
variables getting out of sync are not really nice to debug.