Handle objects with no direct PNG for CEGUI in editor.
Some special game objects do not have a direct PNG image set as
graphic. For example, the thromp's graphic is set in level_itemd.xml to
up_1.png, which does not exist. There is only up_1.settings, which uses
the .settings file's "base" command to refer to the actual PNG
up.png. CEGUI does not know how to handle .settings files as they're a
TSC-specific. However, TSC did not provide a way to find out the PNG
that was actually loaded before this commit; to resolve that problem,
this commit adds cGL_Surface::Get_Real_PNG_Path() whose return value is
that one actually loaded PNG. Determining that required some plumbing in
TSC's lowlevel image loading guts, though.

Other objects do not have an image set at all, most notably the path
object. For those, the generic dummy image is shown currently, but
another image might be more appropriate as the image is not exactly
broken, it does not exist by defintion.

This commit reverts the edits in the level_items.xml file, the editor is
now able to process all items for graphics.