Give the editors their own sprite manager.
For each object that can be placed in the game, the editor must hold
a template object from which the final object is then copied (by
means of its Copy() method). Before this commit that template object
was stored inside the active level's sprite manager. That was
unfortunate, because the editors are instanciated when the game starts
up. That it worked was only an incidiential result of the objects being
stored in the menu level's sprite manager. When trying to also
instanciate the world editor (which is commented out) this crashed the
game with a segfault, because there was no active overworld on game
startup (there is no such thing as a menu world).

With this commit, the editor objects now have their own sprite manager
to store these template objects, making them independant of any menu
level.

Following are two commits that will adapt the level and world editor
sources accordingly.