Remove inappropriate FIXMEs.
Due to the nature of the editor it is not possible to place legacy XML
tags in the level. The level_items.xml and world_items.xml files should
only contain non-legacy content, and inserting anything else is not
possible with the editor. Thus, while loading the editor menu, it is not
required to cater for legacy elements (level loading of course is
unaffected by this commit).

This commit goes as far and causes the code to throw an exception if a
legacy tag is detected in the editor items XML file. This is to make the
developer aware that this file needs to be updated when a multi-sprite
legacy object is introduced.

Also cf. the more lengthy explanation this commit adds to
editor_items_loader.cpp.