Re-enable the mover mode in editor.
The implementation is slightly different from that one in the old TSC
editor, which directly forwarded the mouse move event from the main
event loop to whatever editor was active. Now the event is processed
just the same way as all other events are: They are handed to the
respective input handler (cMouseCursor in this case), which forwards
them to the correct target (the active editor). There is no reason to
treat the mouse move event different from other events.