Add a second panel to the editor.
This panel will be used for editing objects inside the level. The
current approach to this has a fairly complex layouting code with
appending new rows or not, and manually calculating widths of widgets,
it currently does not work in the new editor (widths would have to be
recalculated), and from a UI perspective, it isn't nice to use (remember
the large panel when editing a particle emitter?). Moving all
configuration widgets into a sidebar on the right side of the editor
that is normally invisible will resolve all of these problems at once.

This commit only installs the panel, but does not change anything else yet.