Fix runtime linker not finding the CEGUI libraries.
CMake drops its libraries into the rather unusual /usr/lib64/cegui-0.8/
directory, i.e. not directly into a standard directory. Since our
FindCEGUI module is smart enough to search such pathes, the libraries
are found for compilation, but the runtime linker isn't as smart and
complained it can't find the shared CEGUI libraries. This commit
instructs CMake to include any directories it needed for linking into the
list of additional runtime search pathes of the TSC executable, which
the runtime linker will consult before falling back to the system's
default directories.

As a result, the problem is now gone on the system that compiled TSC
originally as there is no in the location of CEGUI between compilation
and run time. If you run TSC on another system then you compiled it on,
the problem might again be exhibited, but in that case it is up to the
user to fix it. We as the devs cannot cater for all possible locations
of shared libraries everywhere. If you are a packager, be sure to set
the runtime linker path accordingly.