Overhaul build system.
The CMakeLists.txt is much cleaner now.

This commit separates out the instructions to build tinyclipboard and
mruby (and prepares doing so for CEGUI, cf. ticket #531). It further
cleans up the entire structure of the file.

The most notable user-visible change is that the FIXED_DATA_DIR option
has been removed and it has been entirely replaced by use of CMake's
standard module GNUInstallDirs, which makes the a number of new path
options available (see the updated INSTALL.md file). This change is
important for packagers (/cc @akien-mga @muammar).

The option USE_SYSTEM_TINYCLIPBOARD has been documented properly.

A new summary output has been added to the end of the configuration
output.

There will be further changes with regard to how compilation for Windows
is executed.