Work on Windows crossbuild.
This commit prepares the build system further for the Win32 crossbuild.

It removes the FindSFML cmake module, because that one didn't deem it
necessary to check pkg-config at all for SFML, which makes it fail with
MXE's appearently nonstandard SFML location. Bare pkg-config works as
expected on the other hand.

It adds a FindRuby module, because cmake's own FindRuby wants to find
the Ruby library as well and causes the build system to abort if the
Ruby library is not found. We however don't need the Ruby library, but
only the Ruby interpreter for running our own and mruby's build scripts.