Optimise jewel boxes.
They called cImage_Manager::Get_Pointer(), which is a very expensive
operation because it does lots of string comparisons, three times.

1. In cGoldpiece::Init() when the object is constructed.
2. In the cJGoldpiece constructor to set the default color of a jumping
   goldpiece.
3. In cBonusBox::Activate() after the object was instanciated to define
   the gold color.

Each of these calls was indirectly hidden by a call to
cGoldpiece::Set_Gold_Color(). This commit adds a value for the gold
color to the constructor of cGoldpiece and its subclasses, thereby
reducing the number of calls to cImage_Manager::Get_Pointer() to 1,
which results in a significantly more fluent experience on jewel boxes.