Optimise cImage_Manager::Get_Pointer().
This commit replaces the lenghtly query-them-all loop in
cImage_Manager::Get_Pointer() with a hash table using the
std::unordered_map class as a key mapping for indices in the objects
array, eliminating the need to loop in cImage_Manager::Get_Pointer()
entirely.

As a result, this commit is an *extreme* performance boost. On my
machine, it just doubled TSC's framerate and made level loading
lightning fast.

It could be made even more performant by

a) Decoupling cImage_Manager from cObject_Manager and having it use
   std::unordered_map directly to map to the GL_Surface instances instead
   of indices in the `objects' vector, i.e. doing away with the `objects'
   vector alltogether, and

b) Somehow getting around the need to use path_to_utf8(). It appears
   boost::filesystem::path instances cannot be used as keys in
   std::unordered_map, it causes compilation errors.

Fixes #550.

Really, if you did not yet run a recent devel snapshot of TSC, try
it. Then apply this commit and compare the performance! It's a huge
difference!