Extract Lubuntu 16.10 instructions into separate file under docs/.
INSTALL.md provides a generic overview of the build process that is
distribution-agnostic. However, it is good to have a concise listing of
commands required to get TSC up and running on a certain
distribution. These are probably best separated into a separate file to
not enlarge the already large INSTALL.md too much.

Keeping the detailed instruction in INSTALL.md also caused problems with
the logical structure of the file, which first explans the dependencies,
then the compilation options, then the compilation from tarball and from
Git, and then crosscompilation. The entire listing for compiling on
Lubuntu 16.10 does not fit that structure as a whole and would have
needed to be broken up (and in fact, was in INSTALL.md where
applicable). Since the full listing in order has value on its own, it is
best maintained in a separate file as well.

It has been placed under docs/pages, where all miscellaneous
documentation resides, and a notice has been added to INSTALL.md that
detailed instructions are available in that file and where it is to be
found.