Add CEGUI XMLSchema validation files.
On Windows, we're currently forced to use the Xerces XML parser, because
that's the parser MXE's CEGUI is configured for. The CEGUI XMLParser
based on Xerces however wants to validate the XML files it parses
against XSD schema files. This commit properly configures the CEGUI
resource group for the XSD schemas and adds the schemas themselvse (they
have been copied from the CEGUI 0.8.7 source tarball).

The Xerces parser is not used for parsing TSC's own XML files, there we
still use libxml++. As of now this means that on Windows TSC has 2 XML
parsers linked in (libxml++ and Xerces).

On Linux, TSC only needs libxml++ and that one is used with CEGUI as
well.

With this commit, TSC crosscompiles for Win32 and starts. It however
segfaults when initialising the image cache.