Add retrieval of rescue item.
This commit adds a new class cRescueItem that represents the falling
rescue item. It replaces the external update code of the old HUD that
basically took a bare cMovingSprite and defined its behaviour from the
outside inside the HUD's Update() method. Now all handling of the rescue
item's interaction with the environment is encapsulated in the
cRescueItem class.