Fix reading of joy_axis_threshold XML element in preferences, fixes #573
The joy_axis_threshold value was incorrectly read as an int rather than a float, causing it to be truncated to zero.  For some reason this also was causing the game to get stuck waiting for a focus event, freezing the game - we may want to consider investigating this further.