Consider the new Package tab in keyboard navigation. Fixes #578.
When the new Package tab was introduced, it was overlooked to adapt the
keyboard navigation for the Start menu accordingly. As a result, CEGUI
was told to focus widgets on the wrong tab, which fails to work. This
commit adds the missing code for the Package tab, rectifying which
widgets get focused, and thereby fixing the problem described in #578.