Improve level exit/entry indicators and convert them to CEGUI.
This commit removes the text indicators next to level exits and entries
as they were before (which were drawn using raw SFML) and replaces them
with a more powerful indicator that is shown next to the X/Y/UID info
when hovering over the level exit resp. level entry. It now not only
shows the name or target entry, but also the level the entry is in if it
differs from the current level. The final exit in a level is designated
with a "FINISH" indicator.