Do not ship mruby-print gem, replace __printstr__ instead.
This commit deletes the file tsc/data/scripting/core/print.rb, which was
copied from mruby earlier completely. To map #puts et al. to the game
console, this commit instead takes advantage of Ruby's dynamic nature
and specifically overwrites the Kernel#__printstr__ method used by the
mruby-print MRBGEM. That way, the mruby-print MRBGEM can now be used
directly and does not have to be copied into TSC; the code simply
exchanges the Kernel#__printstr__ method provided by it with a custom
implementation that outputs to the game console.

The removal of the named file removes an externally licensed (MIT) file
from TSC own code base.

As a side effect, the code loading the scripts in the
data/scripting/core directory automatically on interpreter startup could
be removed as well.