Do not run player moving code in Update_Waking() if game console is open.
cLevel_Player::Update_Walking() assumed that if the left or right arrow
key is pressed, the player is moving. This assumption does not hold true
any longer if the game console is open. As a result, the player started
a weird move if the left or right arrow key was used with the console
open.

This commit fixes the problem by fixing the assumption in
cLevel_Player::Update_Walking().