Handle Window Close event also if not focused. Fixes #517.
Hitting ALT+F4 on a number of systems causes the respective window
manager to catch it, thus it never reaches TSC's event loop at all as
a keypress. Instead, the window manager generates a WM_CLOSE event and
submits that one to the game loop, or at least it should. At least the
i3 window manager previously generates a WM_UNFOCUS event first before
it immediately afterwards generates WM_CLOSE. As soon as TSC's game
loop encounters a Window Unfocus event however, it enters a special
wait mode that only waited for the corresponding Window Focus event
and ignored all other events that might get passed to the event
handler, including a Window Close event.

This commit changes the behaviour to also cater for Window Close
events submitted to the game loop in the unfocused mode. As a result,
this commit not only fixes ticket #517, but also fixes the previously
undiscovered problem that TSC could not be closed while the window is
not focused.

In #517 it was described that the problem does not appear for
everyone. The reason for this is that window managers may directly
generate the WM_CLOSE event, in which case the event loop always
properly processed it as a request for game quit. The problem only
appeared with window managers that unfocused the window (like i3)
before submitting the Close event.

(terminology notice) WM_CLOSE and WM_UNFOCUS are X11 event terms. They
get transformed by SFML into Window Close and Window Unfocus events,
respectively. Win32 uses other terms, but SFML maps things uniformly
to Window Close and Window Unfocus to ease application developer's
life.