Add secret area trigger.
A secret area is triggered when the user presses a movement key inside
it. This is suboptimal, but consistent with the rest of TSC's
invisible object "colliders", e.g. the level exit. It will not detect
a secret area if the user walks through it without ever
releasing+pressing a key (likewise, a level exit would not activate).