Add selector menu and allow dragging of secret areas.
This is buggy with regard to the start position of an area (which
appears to be forcibly set to (0|0)). Next step is to fix that.

The selector menu added with this commit is a new tab in the left
panel of the editor that controls what happens when the user
leftclicks and drags. The "Selection" option reflects the default
behaviour as it always has been: it causes object selection. The other
option, "Secret Area", allows to add a secret area by dragging
instead.