Improve secret area activation detection.
Instead of tacking it to the player input, simply check for a player
intersection in cSecret_Area::Update() if the secret area is in the
camera range. This ensures that even if the player falls into a secret
area from above without pressing a key that the area is activated
properly.