Move Process_Input() call to cEditor::Update().
This commit fixes the arrow keys not working in the world editor
because the world editor's Process_Input() function was never
called. For whatever reason, the call to that function was in
cLevel_Manager::Update(), where it was only conducted for the level
editor (obviously). By moving it to cEditor::Update() it now gets
called for all editors.