Add player "Die" event and Player#veto_death. Part of #285.
As has been requested in #285, this commit implements a "Die" event for
the player and allows the script writer to veto the player's death in
a specific case by calling Player#veto_death from within the "Die"
event's handler. It is required that the script writer implements a
useful behaviour to happen instead, otherwise Alex may get stuck on an
enemy or fall down eternally after crossing the lower level edge.