Change credits_item_clicked() handler to return bool.
The CEGUI subscriber slot overload allowing a `void` return value
appeared first in 0.8.5, making it impossible to compile TSC against
CEGUI 0.8.4, which is pretty common still. As the `void` overload is
not specifically needed, this commit changes this to allow such
compilation.

This commit is intended to fix the problem described in ticket #593.