Fix compilation warning on sprintf() use.
Appearently the copyright update script messed with the copyright
notice output for the game console. This should not happen again,
since the game_console.cpp file now has a copyright header as it
should (since the last commit), so the script will not replace the
copyright notice in the code again (for each file, it only searches
for the first "Copyright" line).