Update all collision rects when entering the editor.
This is required so that the initial rect's values are set properly,
which in turn are checked when checking for collisions of the mouse
cursor with the objects the user wants to edit. Prior to this commit,
it could happen that the editor did not realise that there is an
object below the cursor, because the rect was where the object
currently was in the level. This was most appearent with moving
enemies: if the enemy had moved away from its initial position, the
user couldn't select it in the editor by hovering over it on the
initial position, where it was drawn. Instead, he had to hover over
the place where the enemy was currently in the level, where it was not
drawn.

The problem is fixed by this commit by executing a position update on
all objects in the current level/world on editor loading.

Part of #597.