Don't use the dummy image for "special"-tagged objects in editor.
The dummy image is reserved for cases where an image is expected, but
not found. For objects tagged with "special" in the editor's item XML
file, there is intentionally no image, and that's what the "special
object" image is for that is reinstated by this commit (it went out of
use when the use of internal CEGUI APIs was removed, confusing users).

This is part of #597.