Fix CMake build issue caused by renaming of license file
Commit id 8a8ac97a0880391f6134480aa6b6df190c461a1f moved and renamed the license file from tsc/docs/license.txt to COPYING (at the top of the directory structure).  The definition of CPACK_RESOURCE_FILE_LICENSE needed to reflect this.  We may want to define a new directory variable so that we do not have to use .. in the path for the COPYING file.