Fix #549: Fading sounds/music no longer hang the game
Previously, fading sounds/music (accidentally) would hang the entire game loop. Obviously, this is very bad.

Now, the fade state is saved within the cAudio engine. When a fade for music is initiated,
the start time, duration, and fade direction are saved, and then the volume is updated at
each iteration of the main loop.

Fading for sounds has been removed, because it would have been harder to implement and no one
could hear the difference anyway...