Place an empty block in new levels. Cf. #565.
This commit reduces the chance that the issue described in ticket #565
appears by placing a simple massive block into new levels. That block
can also be seen as default level content you are free to remove.

The problem isn't really remedied, though. If you delete the default
block, place a lava object in the level, save, then quit and load the
level again, the lava is invisible again.