Update minipoints and text messages not only in level mode. Fixes #605.
The minipoints shown e.g. when collecting jewels or hitting enemies
did not disappear when switching to the world map or any other mode
other than level mode, because the minipoints update was restricted to
level mode. This commit also fixes the related problem that the
same was true for HUD messages, like "Omega mode enabled" or
"Screenshot saved".