Support both libxml++2.6 and libxml++3.0.
This commit introduces a new compilation option USE_LIBXMLPP3. If this
option is enabled (it is disabled by default), then the build system
will search for libxml++3.0 instead of libxml++2.6, and the code uses
the newer libxml++3.0 API instead as well.

Enabling this option is marked EXPERIMENTAL, because the current LTS
distributions Debian 9 and Ubuntu 18.04 do not ship with libxml++3.0
and thus can't be tested on.

To use it, invoke cmake with -DUSE_LIBXMLPP3=ON.