Override image set setting in cShell::Set_Color(). Fixes #599.
cShell's XML constructor calls cShell::Set_Color() to set the shell's
colour according to what was found in the XML. Since the default
colour for armys (and shells are armys) is red, the delegated
cArmy::Set_Color() does an early return with doing nothing as it
recognised the colour is already red if a red shell was requested. If
however a green shell was requested, cArmy::Set_Color() changes the
colour accordingly, and as the last operation, sets the current image
set to army's walking image set. Thus, the problem described in ticket

The problem is fixed by this commit by overriding the active image set
to the shell image set in the cShell's specialised cShell::Set_Color()
after cArmy::Set_Color() has been called. Setting ARMY_SHELL_STAND as
the moving state is not required, because that one is set already in
cShell::Init() and is not touched by Set_Color().