Restructure scripting timers.
In order to fix the problem described in ticket #118, it was necessary
to restructure the timer thread's thread function. With the changes
made by this commit, it becomes now possible to set pause a
timer. Pausing effectively freezes the ticking counter, i.e. once the
timer is continued, it will continue ticking where it left off.

The new functionality isn't yet used, this will be the subject of the
next commit.

The internal restructuration caused the behaviour of the scripting
APIs Timer#stop, Timer#stop! and Timer#interrupt to
change. Timer#stop! and its alias Timer#interrupt are
removed. Timer#stop is still a blocking method, but that doesn't hurt
anymore because the stopping condition is now checked several times a
second, i.e. is executed quickly upon. Thus, Timer#stop now doesn't
wait anymore for the next iteration of the timer; it quickly
aborts. Thereby, the execution of another iteration of the timer's
callback as it was the case previously with Timer#stop has been
cancelled (that is, calling Timer#stop makes the timer stop without
calling the callback again now).