Comment mp_quit_active.
This variable is meant to show the quit icon next to the Quit start
menu entry, but there is no quit icon yet -- it has not yet been
created. So, instead of trying to load a nonexisting file, simply do
nothing. The commented code shows where the icon would go if it gets
made at some point.