Kill enemies if the standing player is invincible. Fixes #373.
If the player does not move, then the collision is only fired on the
enemy, but not on the player. Thus, it's not sufficient if the
player's collision handler kills enemies, resulting in the effect
described in ticket #373. This commit resolves the problem by checking
the player's invincibility also in the enemies' collision handlers.